requirements.txt
azure-ai-formrecognizer==3.3.0
azure-core>=1.30.0
pandas==2.1.4
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
Pillow==10.1.0
openai>=1.6.1